requirements.txt
pymongo
dnspython
pyyaml
python-dateutil